Add gameover screen with statistics

+++ Assets/Scripts/Types/Interfaces.cs
@@ -30,6 +30,6 @@ namespace Landkreuzer.Types {
 	}
 
 	public interface IDamaging {
-		int Damage { get; }
+		float Damage { get; }
 	}
 }
